Read commit messages through cached store
This removes an unnecessary double-read. There are still a few `clone`s
of `Commit` values to prevent overlapping mutable borrows, which isn’t
excellent and is sometimes conceptually unnecessary (when the second
borrow only uses functionality of the `GitStore` that doesn’t actually
overlap with the first-borrowed content). Still, it’s much cheaper than
spawning the extra subprocesses.