Include branch directives in `[update]` messages
Like `[update patch: some-directive]`; likewise for diffbase. This
should be safe in terms of not duplicating cross-references, and is at
least not _less_ safe than currently since the branch directives already
appear in a trailer.

The intent is to make longer chains more readable and help reviewers
more quickly identify the salient changes in the GitHub UI. We will see
how well it fares.